+++ example/lib/main.dart
@@ -20,21 +20,20 @@ class App extends StatelessWidget {
     routes: [
       GoRoute(
         path: '/',
-        builder: (context, state) => MaterialPage<void>(
+        pageBuilder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: const Page1Page(),
         ),
       ),
       GoRoute(
         path: '/page2',
-        builder: (context, state) => MaterialPage<void>(
+        pageBuilder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: const Page2Page(),
         ),
       ),
     ],
-
-    error: (context, state) => MaterialPage<void>(
+    errorPageBuilder: (context, state) => MaterialPage<void>(
       key: state.pageKey,
       child: ErrorPage(state.error),
     ),