+++ example/lib/named_routes.dart
@@ -26,7 +26,7 @@ class App extends StatelessWidget {
       GoRoute(
         name: 'home',
         path: '/',
-        builder: (context, state) => MaterialPage<void>(
+        pageBuilder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: HomePage(families: Families.data),
         ),
@@ -34,7 +34,7 @@ class App extends StatelessWidget {
           GoRoute(
             name: 'family',
             path: 'family/:fid',
-            builder: (context, state) {
+            pageBuilder: (context, state) {
               final family = Families.family(state.params['fid']!);
               return MaterialPage<void>(
                 key: state.pageKey,
@@ -45,7 +45,7 @@ class App extends StatelessWidget {
               GoRoute(
                 name: 'person',
                 path: 'person/:pid',
-                builder: (context, state) {
+                pageBuilder: (context, state) {
                   final family = Families.family(state.params['fid']!);
                   final person = family.person(state.params['pid']!);
                   return MaterialPage<void>(
@@ -61,7 +61,7 @@ class App extends StatelessWidget {
       GoRoute(
         name: 'login',
         path: '/login',
-        builder: (context, state) => MaterialPage<void>(
+        pageBuilder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           // pass the original location to the LoginPage (if there is one)
           child: LoginPage(from: state.params['from']),
@@ -69,7 +69,7 @@ class App extends StatelessWidget {
       ),
     ],
 
-    error: (context, state) => MaterialPage<void>(
+    errorPageBuilder: (context, state) => MaterialPage<void>(
       key: state.pageKey,
       child: ErrorPage(state.error),
     ),