+++ example/lib/redirection.dart
@@ -28,14 +28,14 @@ class App extends StatelessWidget {
     routes: [
       GoRoute(
         path: '/',
-        builder: (context, state) => MaterialPage<void>(
+        pageBuilder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: HomePage(families: Families.data),
         ),
         routes: [
           GoRoute(
             path: 'family/:fid',
-            builder: (context, state) {
+            pageBuilder: (context, state) {
               final family = Families.family(state.params['fid']!);
               return MaterialPage<void>(
                 key: state.pageKey,
@@ -45,7 +45,7 @@ class App extends StatelessWidget {
             routes: [
               GoRoute(
                 path: 'person/:pid',
-                builder: (context, state) {
+                pageBuilder: (context, state) {
                   final family = Families.family(state.params['fid']!);
                   final person = family.person(state.params['pid']!);
                   return MaterialPage<void>(
@@ -60,14 +60,14 @@ class App extends StatelessWidget {
       ),
       GoRoute(
         path: '/login',
-        builder: (context, state) => MaterialPage<void>(
+        pageBuilder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: const LoginPage(),
         ),
       ),
     ],
 
-    error: (context, state) => MaterialPage<void>(
+    errorPageBuilder: (context, state) => MaterialPage<void>(
       key: state.pageKey,
       child: ErrorPage(state.error),
     ),