+++ example/lib/sub_routes.dart
@@ -21,14 +21,14 @@ class App extends StatelessWidget {
     routes: [
       GoRoute(
         path: '/',
-        builder: (context, state) => MaterialPage<void>(
+        pageBuilder: (context, state) => MaterialPage<void>(
           key: state.pageKey,
           child: HomePage(families: Families.data),
         ),
         routes: [
           GoRoute(
             path: 'family/:fid',
-            builder: (context, state) {
+            pageBuilder: (context, state) {
               final family = Families.family(state.params['fid']!);
 
               return MaterialPage<void>(
@@ -39,7 +39,7 @@ class App extends StatelessWidget {
             routes: [
               GoRoute(
                 path: 'person/:pid',
-                builder: (context, state) {
+                pageBuilder: (context, state) {
                   final family = Families.family(state.params['fid']!);
                   final person = family.person(state.params['pid']!);
 
@@ -54,7 +54,7 @@ class App extends StatelessWidget {
         ],
       ),
     ],
-    error: (context, state) => MaterialPage<void>(
+    errorPageBuilder: (context, state) => MaterialPage<void>(
       key: state.pageKey,
       child: ErrorPage(state.error),
     ),